+++ kernel/shell/history/history.cpp
@@ -5,11 +5,11 @@
 #include "print.h"
 #include "line_editor.h"
 
-char commandHistory[MAX_COMMAND_HISTORY][TERMINAL_BUFFER_SIZE];
+char commandHistory[MAX_COMMAND_HISTORY][LINE_EDITOR_BUFFER_SIZE];
 uint64_t cmdHistCount = 0;
 uint64_t goThroughHistoryCount = 0;
 
-void add_command_to_history(char command[TERMINAL_BUFFER_SIZE]) {
+void add_command_to_history(char command[LINE_EDITOR_BUFFER_SIZE]) {
     if (cmdHistCount >= MAX_COMMAND_HISTORY) {
         for (uint64_t i = 0; i < cmdHistCount - 1; i++) {
             str_copy(commandHistory[i], commandHistory[i + 1]);